+++ AivenEcommerce.V1.Domain.Shared/Dtos/CouponCodes/CreateCouponCodeInput.cs
@@ -9,7 +9,7 @@ namespace AivenEcommerce.V1.Domain.Shared.Dtos.CouponCodes
     public record CreateCouponCodeInput(
         string Code,
         CouponCodeOffType Type,
-        decimal Value,
+        int Value,
         int MinAmount,
         int? MaxAmount,
         IEnumerable<string> Categories,
