+++ AivenEcommerce.V1.Domain.Shared/Dtos/Products/UpdateProductCostPriceInput.cs
@@ -1,4 +1,4 @@
 ﻿namespace AivenEcommerce.V1.Domain.Shared.Dtos.Products
 {
-    public record UpdateProductCostPriceInput(string ProductId, decimal Cost, decimal Price);
+    public record UpdateProductCostPriceInput(string ProductId, decimal Cost, int Price);
 }
