+++ AivenEcommerce.V1.Domain.Shared/Dtos/Products/UpdateProductInput.cs
@@ -6,7 +6,7 @@ namespace AivenEcommerce.V1.Domain.Shared.Dtos.Products
     public record UpdateProductInput(string Id,
         string Name,
         decimal Cost,
-        decimal Price,
+        int Price,
         short PercentageOff,
         string Category,
         string SubCategory,
